Split runner profile process lifecycle tests
